+++ schedule/schedule-appengine/src/main/client/jasify/directives/email.directive.js
@@ -1,0 +1,23 @@
+(function (angular) {
+
+    angular.module('jasify').directive('jasEmail', jasEmail);
+
+    function jasEmail($q, Unique) {
+        return {
+            require: 'ngModel',
+            link: function (scope, elm, attrs, ctrl) {
+
+                ctrl.$asyncValidators.jasEmail = function (modelValue, viewValue) {
+
+                    if (ctrl.$isEmpty(modelValue)) {
+                        return $q.when();
+                    }
+
+                    var def = $q.defer();
+
+                    return Unique.email(modelValue);
+                };
+            }
+        };
+    }
+})(angular);