+++ schedule/schedule-appengine/src/main/client/jasify/templates.module.js
@@ -1,0 +1,5 @@
+(function (angular) {
+
+    angular.module('templates', []);
+
+})(angular);