+++ schedule/schedule-appengine/src/test/java/com/jasify/schedule/appengine/http/json/JsonUserTest.java
@@ -5,7 +5,6 @@ import com.jasify.schedule.appengine.model.common.Organization;
 import com.jasify.schedule.appengine.model.users.User;
 import com.jasify.schedule.appengine.util.JSON;
 import com.jasify.schedule.appengine.util.TypeUtil;
-import junit.framework.Assert;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;