+++ schedule/schedule-appengine/src/test/java/com/jasify/schedule/appengine/spi/dm/JasUserListTest.java
@@ -6,7 +6,7 @@ import org.junit.Test;
 import java.util.ArrayList;
 import java.util.List;
 
-import static junit.framework.Assert.assertEquals;
+import static junit.framework.TestCase.assertEquals;
 
 public class JasUserListTest {
     @Test