+++ schedule/schedule-appengine/src/test/java/com/jasify/schedule/appengine/validators/UsernameValidatorTest.java
@@ -1,7 +1,7 @@
 package com.jasify.schedule.appengine.validators;
 
 import com.jasify.schedule.appengine.TestHelper;
-import com.jasify.schedule.appengine.model.users.User;
+import com.jasify.schedule.appengine.model.users.TestUserServiceFactory;
 import com.jasify.schedule.appengine.model.users.UserServiceFactory;
 import org.junit.After;
 import org.junit.Before;
@@ -9,9 +9,13 @@ import org.junit.Test;
 
 import java.util.List;
 
-import static junit.framework.TestCase.*;
+import static junit.framework.TestCase.assertNotNull;
+import static junit.framework.TestCase.assertTrue;
+import static org.easymock.EasyMock.expect;
 
 public class UsernameValidatorTest {
+    private TestUserServiceFactory testUserServiceFactory = new TestUserServiceFactory();
+
     private UsernameValidator validator;
 
     private static void assertContains(List<String> list, String text) {
@@ -20,41 +24,35 @@ public class UsernameValidatorTest {
 
     @Before
     public void servletRunner() {
-        TestHelper.initializeJasify();
+        TestHelper.initializeDatastore();
         validator = UsernameValidator.INSTANCE;
         assertNotNull(validator);
+        testUserServiceFactory.setUp();
     }
 
     @After
     public void stopDatastore() {
         TestHelper.cleanupDatastore();
+        testUserServiceFactory.tearDown();
     }
 
     @Test
     public void testValidateGood() throws Exception {
-        List<String> list = validator.validate("krico");
-        assertTrue(list.isEmpty());
+        expect(UserServiceFactory.getUserService().usernameExists("krico")).andReturn(false);
+        testUserServiceFactory.replay();
+        assertTrue(validator.validate("krico").isEmpty());
     }
 
     @Test
     public void testValidateUserExists() throws Exception {
-
-        List<String> list = validator.validate("krico");
-        assertTrue(list.isEmpty());
-
-        User user = UserServiceFactory.getUserService().newUser();
-        user.setName("krico");
-        UserServiceFactory.getUserService().create(user, "password");
-        list = validator.validate("krico");
-        assertFalse(list.isEmpty());
-        assertEquals(1, list.size());
-        list = validator.validate("KriCo");
-        assertFalse(list.isEmpty());
-        assertEquals(1, list.size());
+        expect(UserServiceFactory.getUserService().usernameExists("krico")).andReturn(true);
+        testUserServiceFactory.replay();
+        assertContains(validator.validate("krico"), UsernameValidator.REASON_EXISTS);
     }
 
     @Test
     public void testValidateNull() throws Exception {
+        testUserServiceFactory.replay();
         List<String> list = validator.validate(null);
         assertContains(list, UsernameValidator.REASON_LENGTH);
         assertContains(list, UsernameValidator.REASON_VALID_CHARS);
@@ -62,6 +60,7 @@ public class UsernameValidatorTest {
 
     @Test
     public void testValidateEmpty() throws Exception {
+        testUserServiceFactory.replay();
         List<String> list = validator.validate("");
         assertContains(list, UsernameValidator.REASON_LENGTH);
         assertContains(list, UsernameValidator.REASON_VALID_CHARS);
@@ -69,6 +68,8 @@ public class UsernameValidatorTest {
 
     @Test
     public void testValidateShort() throws Exception {
+        expect(UserServiceFactory.getUserService().usernameExists("kri")).andReturn(false);
+        testUserServiceFactory.replay();
         List<String> list = validator.validate("k");
         assertContains(list, UsernameValidator.REASON_LENGTH);
 
@@ -82,6 +83,7 @@ public class UsernameValidatorTest {
 
     @Test
     public void testValidateCharacters() throws Exception {
+        testUserServiceFactory.replay();
         List<String> list = validator.validate("kri@co");
         assertContains(list, UsernameValidator.REASON_VALID_CHARS);